Modified Changes for Mac OS Sequoia 15.5

+++ asitop_csv_logger/utils.py
@@ -3,7 +3,7 @@ import glob
 import subprocess
 from subprocess import PIPE
 import psutil
-from parsers import *
+from .parsers import *
 import plistlib
 
 
@@ -48,7 +48,12 @@ def run_powermetrics_process(timecode, nice=10, interval=1000):
     #ver, *_ = platform.mac_ver()
     #major_ver = int(ver.split(".")[0])
     for tmpf in glob.glob("/tmp/asitop_powermetrics*"):
-        os.remove(tmpf)
+        try:
+            if os.access(tmpf, os.W_OK):
+                os.remove(tmpf)
+        except Exception:
+            pass
+
     output_file_flag = "-o"
     command = " ".join([
         "sudo nice -n",
